+++ server/src/test/java/org/realityforge/replicant/server/TestSession.java
@@ -1,9 +1,7 @@
 package org.realityforge.replicant.server;
 
 import javax.annotation.Nonnull;
-import javax.json.stream.JsonGenerator;
 import org.realityforge.replicant.server.transport.ReplicantSession;
-import org.realityforge.rest.field_filter.FieldFilter;
 
 public class TestSession
   extends ReplicantSession
@@ -11,10 +9,5 @@ public class TestSession
   public TestSession( @Nonnull final String sessionID )
   {
     super( sessionID );
-  }
-
-  @Override
-  public void emitStatus( @Nonnull final JsonGenerator g, @Nonnull final FieldFilter filter )
-  {
   }
 }