+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 0.5.64 (Pending):
+* Move the conversion of sessions to json for rest service from ReplicantSession to EncoderUtil
+  class in same package as AbstractSessionRestService.
 * Introduce `JsonUtil` as a helper to convert strings and structured types to `JsonObject`
   instances.
 * Make `ChannelDescriptor` implement `Comparable` to enable sorting.